requirements.txt
Django==5.0.2
django-allauth==0.61.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
hdbscan==0.8.33
numpy==1.26.3
pandas==2.1.4
pydantic==1.10.14
scikit-learn==1.4.0
zenml